requirements.txt
requests
python-dotenv
twitchio
rcon
websockets
gtts
pydub
